Adding support for M1-M3 apple silicon chips
Modified the main.tf and the ec2.tf to allow for support of the newest M1 through M3 chips from Apple. This was made to get around the problem with template from hashicorp's deprecation of further development of the M1 branch of template. Version v2.2.0 is no longer supported on darwin-amd64.